Updates Class Structure
Updates the Class structure to more closely align with the `wisper-sidekiq` gem.
