+++ lib/wisper/resque.rb
@@ -2,16 +2,20 @@ require 'wisper'
 require 'resque'
 
 require 'wisper/resque/version'
-require 'wisper/resque/listener'
-
-require 'wisper/broadcasters/resque_broadcaster'
 
 module Wisper
-  module Resque
-    def self.setup
-      Broadcasters::ResqueBroadcaster.register
+  class ResqueBroadcaster
+    def broadcast(listener, _publisher, event, args)
+      ::Resque.enqueue(listener, event, *args)
+    end
+
+    def self.register
+      Wisper.configure do |config|
+        config.broadcaster :resque, ResqueBroadcaster.new
+        config.broadcaster :async,  ResqueBroadcaster.new
+      end
     end
   end
 end
 
-Wisper::Resque.setup
+Wisper::ResqueBroadcaster.register
