--- lib/wisper/resque/listener.rb
@@ -1,9 +1,0 @@
-module Wisper
-  module Resque
-    module Listener
-      def perform(event, *args)
-        new.public_send event, *args
-      end
-    end
-  end
-end
