+++ spec/spec_helper.rb
@@ -9,9 +9,8 @@ RSpec.configure do |config|
     mocks.verify_partial_doubles = true
   end
 
-  config.shared_context_metadata_behavior = :apply_to_host_groups
-
-  config.filter_run_when_matching :focus
+  config.filter_run :focus
+  config.run_all_when_everything_filtered = true
 
   config.disable_monkey_patching!
